requirements.txt
Flask==0.10.1
Flask-Assets==0.10
Flask-Script==2.0.5
Flask-WTF==0.10.0
Jinja2==2.7.3
MarkupSafe==0.23
MySQL-python==1.2.5
SQLAlchemy==0.9.6
WTForms==2.0.1
Werkzeug==0.10.1
bcrypt==1.1.0
cffi==0.9.1
cssmin==0.2.0
distribute==0.6.28
itsdangerous==0.24
pycparser==2.10
pycrypto==2.6.1
six==1.9.0
ujson==1.33
webassets==0.10.1
wsgiref==0.1.2
